feat(OLM-952): support importing proto files from go modules
These changes allow settings service to move forward with leveraging
settings payloads defined in the `schema` repo directly in the settings
service.

Also of note: This only account for generating the pb code within
the repo itself. It won't generate the imported code, that step needs to
be done manually for the service and it's clients

Generation tested manually with and without imports

Example:
```
go-protoc-imports:
  - "github.com/org/repo@latest"
```