feat: Add 'go.sh' to run the current project's Go
Adds a 'go.sh' script that attempts to infer which is the correct
version of Go for this given project then run it.

Normally `asdf` magically takes care of this by looking at your current
directory and the local `tool-versions` file.  But there are situations
where that doesn't work.  For example, a text editor might be started
from a context where `asdf` doesn't apply, or may have a `$PWD` that
doesn't match the project directory.