feat: Add 'asdf-exec.sh' and 'go.sh' (#204)
Adds an `asdf-exec.sh` script that will `asdf exec` the given arguments
in an environment that respects the embedding project's `.tool-versions`
file.

In a perfect world, this functionality would be baked in to `asdf`
upstream.  For now we build this functionality from hacky shell scripts.
It may break in some edge case scenarios, but it should be good enough
for all the situations we're likely to encounter int he real world.

Also adds a `go.sh` that does `asdf-exec.sh go`.  This is intended to
provide compatibilty for tools require us to point at the "Go
executable".

Co-authored-by: Jared Allard <jaredallard@users.noreply.github.com>