fix: DT-3372 Make devconfig env `local_development`
[DT-3372](https://outreach-io.atlassian.net/browse/DT-3372)

Fixes a bug in `devconfig.sh`'s invocation of `build-jsonnet.sh` so it
will render the jsonnets in `local_development` mode instead of
`development` mode.

As I understand it, `local_development` is the environment for
developing on the developer's "host" machine.  The `development`
environment is for deploying to the developer's local k8s instance (ie.
`bento1a`).  The `devconfig` script is only relevant for
`local_development`.

The `devconfig.sh` script made a clear attempt to set the environment to
`local_development`, but it did so via a now-incorrect environment
variable.  In practice, it was rendering jsonnets with an environmetn
value of `development`.

I believe this confusion has been around for roughly six months now.
See ticket for details.  This is unlikely to be a breaking change for
most services, but I am slightly concerned that some services may have
come to depend ont he buggy behavior over this time period.