feat: QSS-1704 Support secrets in devconfig
[QSS-1704](https://outreach-io.atlassian.net/browse/QSS-1704)

Update `devconfig.sh` to support the installation of local overrides for
kubernetes secrets defined in deployment jsonnets.

While `devconfig.sh` has long supported local installation of secrets
defined via the `VaultSecret` CRD, it hasn't supported plain old k8s
secrets until now.

We do not expect this feature to be widely used.  In 99% of cases, a
`VaultSecret` defined in the service's `service.yaml` is the right way
to handle secrets.  However, there are rare use cases, like the one
described in the attached ticket, where one might want to hard-code a
secret in jsonnets and have `devconfig.sh` know how to install it in
`$HOME/.outreach`.