fix: DTSS-1744 Correct vault secret fetching env
[DTSS-1744](https://outreach-io.atlassian.net/browse/DTSS-1744)

Updates the environment override for the `build-jsonnet.sh` invocation
for fetching non-vault secrets to `DEVENV_DEPLOY_ENVIRONMENT`.  This
is analogous to the fix in [1].

The feature to render non-vault secrets in [2] just a few days after [1]
was merged.  However, it seems we made the same mistake as had been
historically present in the config-rendering `build-jsonnet.sh`.  Oops.

[1] https://github.com/getoutreach/devbase/pull/411
[2] https://github.com/getoutreach/devbase/pull/416