fix(e2e): handle armored/dearmored apt key GCP (#544)
* fix(e2e): handle armored/dearmored apt key GCP

Currently, the GCP key is using the wrong file extension for their GPG
key. They have armored it, when that should be a `.asc` file instead.
This causes `apt` to ignore the file silently. We work around this by
dearmoring the file ourself. We also handle if the file potentially gets
fixed one day and no longer needs to be dearmored.

I'm not very attached/satisfied with my grep approach, I'm open to other
ideas (e.g., always attempting to dearmor).

* remove sudo from curl

* small keyring path optimization

* Update shell/ci/testing/setup-devenv.sh

Co-authored-by: Jesse Kinkead <jesse.kinkead@outreach.io>

---------

Co-authored-by: Jesse Kinkead <jesse.kinkead@outreach.io>