fix(e2e): use correct gpg dearmor command
When operating on a file, `gpg --dearmor` will write the output to
another file on disk. However, if you pipe the file it will write to
stdout. I tested this while piping, so this broke once I released it.

Instead we now tell gpg to always output to stdout using `--output -`.
I've tested the entire function in CI now, so I'm confident this works.