feat: support rtx instead of asdf
This PR adds support for rtx[1] instead of the asdf version manager.
This functionality can be disabled with `DEVBASE_RTX_DISABLE_RTX` and is
enabled if `rtx` exists in `PATH`.

We're mostly shimming `rtx` support because it is compatible out of the
box with `asdf`. Longer term we should have hot swappable
implementations using interfaces when we move this functionality to Go.
The main driver of this is a desire to move away from `asdf` longer
term.