fix: better array parsing, small testing framework, asdf list fix (#561)
Fixes the `docker.sh` array parsing to properly handle an array not
being set. Extracts aforementioned logic to be in a generic `yaml.sh`
library in hopes it can be more usable in the future, since moving to Go
isn't happening yet.

As part of that, I felt we needed some tests here. So, I wrote a _very_
simple testing framework to call `*_test.sh` scripts, much like Go. If
the text returns a non-zero exit code, it is considered failed. Added
some tests for the new logic.

I will add some documentation for this in the PR shortly, but I wanted to
get it out for comments (since it is simple and described here).

Fixed an issue where newer asdf versions seem to put `*` in front of
versions as well as a space, so our original `^` regex no longer
works.