fix(docker): properly handle string values again (#564)
Fixes the new `docker.sh` logic to properly support string values again.
Adds a test for the `get_image_field` function and changes `docker.sh`
to support being tested (albeit a little hackily).

Adds a `yaml_get_field` function that strips `null` for an empty string
to be easier to work with in Bash. Adds tests for the new
`yaml_get_field` function and behaviour.

[DT-3821]