feat: rewrite test.sh to be better documented and use bash arrays
Changed `test.sh` to fully document all environment variables in use.
Changed `_FLAG` and `_TAG` environment variables to use bash arrays to
ensure consistent space handling behaviour, while keeping space
splitting for pre-specified environment variables for compatability.

When passing `-v` to `TEST_FLAGS` (and in turn to `go test`) we now set
the `gotestsum` output format to be `standard-verbose` which is
equivalent to `go test -v`[1] output. Before, this was silently ignored
and frustrating to work with.

Updated the `shared/setup_docker` step to use a newer version of Docker
as the older version had issues with apt package installs (somehow?)

[1]: https://github.com/gotestyourself/gotestsum#:~:text=standard%2Dverbose%20%2D%20the%20standard%20go%20test%20%2Dv%20format.