fix: update docker.sh statement, deduplicate logic (#606)
As a follow up to #605, this updates the docker.sh output to better
reflect what the docker build is doing.

I also took the time to combine the logic for the docker build _test_
logic and the docker build _release_ logic (determined by the presence
of `CIRCLE_TAG`. This brings the added benefit of us building a docker
image on all archs that we would release to on PRs.

Updated buildx to reflect what is in use on most user's machines.