feat(golangci-lint): Upgrade to 1.53.3
Upgrades `golangci-lint` to `1.53.3` to support Go 1.20.

This likely brings in some new lint rules that _may_ fail on new
projects. Putting it in unstable to give us sometime to run into them
and adjust our rules as needed.

I didn't hit any in a few repos I tried this on.