fix(golangci-lint): upgrade to v1.54.1, fail on invalid version pairs
Upgrades golangci-lint to v1.54.1 to support Go 1.21. Adds a small
framework in `golangci-lint.sh` to explode if versions of Go and
golangci-lint are not compatible. Adds checks just for 1.21 and 1.20.
This is important because, most notably for 1.20, it can explode in CI
by consuming all the RAM and not be clear why that is happening.

Changes `GOGC` to no longer be set in favor of using total memory
calculation to pass to `GOMEMLIMIT`. Fallsback to `GOGC` if required.
It's important to note that this doesn't disable GC, just makes it less
important until we get close to total memory (or are otherwise unable to
free any).

Also updated `linter.sh`'s error to not blindly suggest `make fmt`.
Instead, it recommends it based on the output.