fix(orbs/save_cache): support setup_remote_docker
Right now, `test` supports `setup_remote_docker` but `save_cache` does
not. This leads to potential issues with tests that need a remote docker
daemon to be available for their tests (thus failing to run on
`save_cache` but _not_ the `test` job).

This, fixes that, by supporting it in `save_cache` as well.