feat(machine): support newer versions of Ubuntu
Adds support for newer versions of Ubuntu and Debian by using the `yq`
apt package, if it's available, and falling back to `pip3` if it's not.

Cleans up the code a bit to make it easier to determine what version of
`gh` is in use/desired. Updates to `v2.32.1`. This will be used by
`bootstrap/dev` in an upcoming PR.