fix(lib/github): import required dependencies on load
While it's not the most performant for us to load dependencies on every
load of the library (since shell functions aren't namespaced), it's more
complicated/problematic to try to make consuming shell scripts aware of
the libary's dependencies. So, we have `lib/github` now import it's own
dependencies.

This could cause some issues with libraries that do initial state
setting (e.g., `asdf`) but most of those libraries should be designed to
only initialize their state once (which `asdf`, the example, does for
this reason)