fix(asdf): use LIB_DIR instead of DIR to avoid clashing
`DIR` is a common pattern used in devbase to determine the current
running directory. Unfortunately, some of our libraries have _also_
started to use it. This ends up in secenarios where `DIR` can be
overwritten by an imported script. To avoid this, we've used `LIB_DIR`
inside the libraries to prevent clashes from causing an issue.

In reality, we shouldn't be using Bash which would help... for now this
fixes it by changing the newly added `DIR` to be `LIB_DIR`.